inventory/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
certifi==2025.1.31
charset-normalizer==3.4.1
drf-nested-routers==0.94.1
idna==3.10
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
